Added the ability to skip the animations in playbacks and storyboards.
We can now skip the entire animation of playbacks by having the user
select an option in the speed control to ‘Skip Animation’. In a normal
playback this makes the tool step to the very end. In a storyboard this
makes the tool skip to the next event that has a comment associated
with it.